Fix __oscap_debuglog_object PID printing
printing PID as signed long is safer

Fix __oscap_debuglog_object PID printing

+++ src/common/debug.c
@@ -209,8 +209,8 @@ void __oscap_debuglog_object (const char *file, const char *fn, size_t line, int
                 if (logfile == NULL)
                         logfile = OSCAP_DEBUG_FILE;
 
-		if (snprintf(pathbuf, sizeof pathbuf, "%s.%u",
-			     logfile, (unsigned int)getpid()) >= (signed int) sizeof pathbuf)
+		if (snprintf(pathbuf, sizeof pathbuf, "%s.%ld",
+			     logfile, (long)getpid()) >= (signed int) sizeof pathbuf)
 		{
                         __UNLOCK_FP;
 			return;
@@ -239,9 +239,9 @@ void __oscap_debuglog_object (const char *file, const char *fn, size_t line, int
 
 #if defined(SEAP_THREAD_SAFE)
         /* XXX: non-portable usage of pthread_t */
-        fprintf (__debuglog_fp, "(%u:%llx) [%s:%zu:%s]\n------ \n", (unsigned int)getpid (), (unsigned long long)pthread_self(), file, line, fn);
+	fprintf (__debuglog_fp, "(%ld:%llx) [%s:%zu:%s]\n------ \n", (long)getpid (), (unsigned long long)pthread_self(), file, line, fn);
 #else
-        fprintf (__debuglog_fp, "(%u) [%s:%zu:%s]\n------\n ", (unsigned int)getpid (), file, line, fn);
+	fprintf (__debuglog_fp, "(%ld) [%s:%zu:%s]\n------\n ", (long)getpid (), file, line, fn);
 #endif
 
         switch (objtype) {
